// src/app/core/models/userInfo.model.ts
export interface UserInfo {
    customer_id: string;
    customer_name: string;
    first_name: string;
    last_name: string;
    customer_short_address: string;
    customer_address_1:string;
    customer_address_2: string;
    city: string;
    state: string;
    country: string;
    zip: string;
    customer_mobile: string;
    customer_email: string;
    image:string;
    password: string;
    company: string;
    status: string;
    error:string
  }
  
  